common: unify proofs_of_indexing/mod.rs & indexing_statuses/mod.rs into queries.rs
