--- backend/crates/common/src/indexing_statuses/mod.rs
@@ -1,2 +1,0 @@
-mod query;
-pub use query::*;
