--- backend/crates/common/src/proofs_of_indexing/mod.rs
@@ -1,3 +1,0 @@
-mod query;
-
-pub use query::*;
