+++ backend/crates/cross-checker/src/tests/indexing_statuses.rs
@@ -1,7 +1,7 @@
 use futures::{future, stream::FuturesUnordered, StreamExt};
 use graphix_common::prelude::IndexingStatus;
 
-use crate::indexing_statuses;
+use crate::queries::query_indexing_statuses;
 
 use graphix_common::tests::{fast_rng, gen::*};
 
@@ -28,11 +28,10 @@ async fn indexing_statuses() {
             .flatten()
             .collect::<Vec<_>>();
 
-        let queried_statuses: Vec<IndexingStatus> =
-            indexing_statuses::query_indexing_statuses(indexers)
-                .await
-                .into_iter()
-                .collect();
+        let queried_statuses: Vec<IndexingStatus> = query_indexing_statuses(indexers)
+            .await
+            .into_iter()
+            .collect();
 
         assert_eq!(expected_statuses, queried_statuses);
     }
